+++ components/table.tsx
@@ -1,27 +1,14 @@
 import Link from "next/link";
 import React from "react";
 import useSWR, { Fetcher } from "swr";
+import Paper from "../types/paper";
 
 
 const Header = () => {
   return (
-    <thead className="bg-gray-50">
-      <tr>
-        <th scope="col" className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-          Title
-        </th>
-        <th scope="col" className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-          Title
-        </th>
-        <th scope="col" className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-          Status
-        </th>
-        <th scope="col" className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-          Conference / Journal
-        </th>
-        <th scope="col" className="relative px-6 py-3">
-          <span className="sr-only">Edit</span>
-        </th>
+    <thead>
+      <tr className="font-normal text-xl text-left">
+        <th>Title</th>
       </tr>
     </thead>
   )
@@ -29,32 +16,24 @@ const Header = () => {
 
 
 const Row = (props: { paper: Paper, key: number }) => {
+  console.log(props.paper);
+
   return (
-    <tbody className="bg-white divide-y divide-gray-200">
+    <tbody className="transitio duration-300 hover:bg-indigo-100">
     <tr>
-      <td className="px-6 py-4 whitespace">
-        <div className="flex items-center">
-          <div className="ml-4">
-            <div className="text-sm font-medium text-sky-700">
-              <Link href={props.paper.paperLink}>{ props.paper.title }</Link>
-            </div>
-            <div className="text-sm text-gray-500">
-              jane.cooper@example.com
-            </div>
-          </div>
+      <td>
+        <div className="text-lg font-mono mx-auto">
+          <Link href={props.paper.paperLink}>
+            <a className="text-indigo-600 hover:text-indigo-900">{ props.paper.title }</a>
+          </Link>
+          <div className="text-sm text-gray-500">{ props.paper.authors }</div>
+          <span className="bg-sky-400 rounded-md">
+          <div className="m-1">{ props.paper.sourceShort }{ props.paper.year }</div>
+        </span>
         </div>
       </td>
-      <td className="px-6 py-4 whitespace-nowrap">
-        <div className="text-sm text-gray-900">Regional Paradigm Technician</div>
-        <div className="text-sm text-gray-500">Optimization</div>
-      </td>
-      <td className="px-6 py-4 whitespace-nowrap">
-        <span className="px-2 inline-flex text-xs leading-5 font-semibold rounded-full bg-green-100 text-green-800">
-          { props.paper.isRead }
-        </span>
-      </td>
-      <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-500">
-        { props.paper.sourceShort }
+      <td className="text-sm flex">
+
       </td>
       <td className="px-6 py-4 whitespace-nowrap text-right text-sm font-medium">
         <a href="#" className="text-indigo-600 hover:text-indigo-900">BibTeX</a>
@@ -62,22 +41,6 @@ const Row = (props: { paper: Paper, key: number }) => {
     </tr>
   </tbody>
   );
-}
-
-
-type Paper = {
-  authors: string;
-  isRead: string;
-  issueLink: string;
-  keywords: string;
-  note: string;
-  paperLink: string;
-  paperType: string;
-  source: string;
-  sourceShort: string;
-  timeStamp: string;
-  title: string;
-  year: string;
 }
 
 
@@ -90,27 +53,12 @@ const Table = () => {
   const isLoading = !error && !data;
 
   return (
-    <div className="flex flex-col">
-      <div className="-my-2 overflow-x-auto sm:-mx-6 lg:-mx-8">
-        <div className="py-2 align-middle inline-block min-w-full sm:px-6 lg:px-8">
-          <div className="shadow overflow-hidden border-b border-gray-200 sm:rounded-lg">
-            { isLoading &&
-              <p>Loading...</p>
-            }
-            { !isLoading &&
-              <table className="min-w-full divide-y divide-gray-200">
-                <Header />
-                {
-                  data?.map((paper: Paper, idx: number) =>
-                    <Row paper={paper} key={idx} />
-                  )
-                }
-              </table>
-            }
-          </div>
-        </div>
-      </div>
-    </div>
+    <table className="table-fixed border-separate border-spacing">
+      <Header />
+      {
+        data?.map((paper: Paper, idx: number) => <Row paper={paper} key={idx} />)
+      }
+    </table>
   );
 }
