+++ pages/interest.tsx
@@ -4,7 +4,7 @@ import Table from "../components/table";
 
 const Interest = () => {
     return (
-        <div className="sm:container mt-48 m-auto flex justify-center">
+        <div className="sm:container mx-auto px-10">
             <Table />
         </div>
     );
