+++ types/paper.ts
@@ -1,0 +1,17 @@
+type Paper = {
+  authors: string;
+  isRead: string;
+  issueLink: string;
+  keywords: string;
+  note: string;
+  paperLink: string;
+  paperType: string;
+  source: string;
+  sourceShort: string;
+  timeStamp: string;
+  title: string;
+  year: string;
+};
+
+
+export default Paper;
